requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.7.2; python_version >= '3.7'

bcrypt==4.0.1

cachetools==5.3.1; python_version >= '3.7'

certifi==2023.7.22; python_version >= '3.6'

cffi==1.15.1

charset-normalizer==3.2.0; python_version >= '3.7'

cloudinary==1.34.0

cryptography==41.0.4; python_version >= '3.7'

defusedxml==0.7.1; python_version >= '3.0'

django-cors-headers==4.2.0

django-jazzmin==2.6.0

django==4.1.11

djangorestframework-jwt==1.11.0

djangorestframework-simplejwt==5.3.0

djangorestframework==3.14.0

djongo==1.3.6

google-auth-httplib2==0.1.1

google-auth-oauthlib==1.1.0

google-auth==2.23.0; python_version >= '3.7'

httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

idna==3.4; python_version >= '3.5'

oauthlib==3.2.2; python_version >= '3.6'

pillow==10.0.1

pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

pycparser==2.21

pyjwt==1.7.1; python_full_version >= '3.7.0'

pymongo==3.12.3

pyparsing==3.1.1; python_version >= '3.1'

python-decouple==3.8

python3-openid==3.2.0; python_version >= '3.0'

pytz==2023.3.post1

requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

requests==2.31.0

rsa==4.9; python_version >= '3.6' and python_version < '4'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

social-auth-app-django==4.0.0

social-auth-core==4.0.2; python_version >= '3.6'

sqlparse==0.2.4; python_version >= '3.5'

typing-extensions==4.8.0; python_version < '3.11'

tzdata==2023.3; sys_platform == 'win32'

urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

